requirements.txt
asgiref==3.11.0
crispy-tailwind==1.0.3
Django==6.0.1
django-crispy-forms==2.5
django-filter==25.2
sqlparse==0.5.5
